More through error message for build script

+++ build.rs
@@ -1,41 +1,60 @@
-use std::fs::File;
-use std::io::Write;
+#![feature(path_ext)]
+
+use std::fs::{File, PathExt};
+use std::io::{stderr, Write};
 use std::process::{Command, Output};
-use std::path;
+use std::path::Path;
 use std::env;
 
-fn run_cmd(name: &str, command: &mut Command) -> Output {
+fn run_cmd(command: &mut Command) -> Output {
+    let stream = stderr();
+    let mut f = stream.lock();
     let res = command.output();
     match res {
         Ok(output) => if !output.status.success() {
-            panic!("{} failed:\n{}", name, String::from_utf8_lossy(&output.stderr[..]))
+            write!(&mut f, "Command `{:?}` failed", command).unwrap();
+            panic!("{}", String::from_utf8_lossy(&output.stderr[..]))
         } else {
             output
         },
-        Err(err) =>
-            panic!("{} failed:\n{}", name, err)
+        Err(err) => {
+            write!(&mut f, "Command `{:?}` failed", command).unwrap();
+            panic!("{}", err)
+        }
     }
 }
 
 fn main()
 {
-    let dir_var = match env::var("OUT_DIR") {
+    let source_file = Path::new("src/codegen/type_size.c");
+    if !source_file.exists() {
+        panic!("Could not find file: {}", source_file.to_string_lossy());
+    }
+
+    let out_dir_var = match env::var("OUT_DIR") {
         Ok(dir) => dir,
         Err(e) => panic!("Could not get env value OUT_DIR: {}", e)
     };
+    let out_dir = Path::new(&out_dir_var);
+    if !out_dir.exists() {
+        panic!("Could not find directory: {}", out_dir.to_string_lossy());
+    }
 
-    let compile_flags = env::var("LIBYAML_CFLAGS").unwrap_or("".to_string());
+    let out_file = out_dir.join("codegen");
 
-    let out_file = path::Path::new(&dir_var).join("codegen");
     let mut gcc_cmd = Command::new("gcc");
-    gcc_cmd.arg("src/codegen/type_size.c")
-        .arg(&compile_flags)
+    gcc_cmd.arg(source_file)
         .arg("-o").arg(&out_file);
 
-    run_cmd("gcc", &mut gcc_cmd);
+    match env::var("LIBYAML_CFLAGS") {
+        Ok(compile_flags) => { gcc_cmd.arg(&compile_flags); },
+        Err(_) => ()
+    }
+
+    run_cmd(&mut gcc_cmd);
 
     let mut codegen_cmd = Command::new(&out_file);
-    let output = run_cmd("codegen", &mut codegen_cmd);
+    let output = run_cmd(&mut codegen_cmd);
 
     let mut f = match File::create("src/type_size.rs") {
         Ok(f) => f,
